# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ByteBurrow API",
    "description": "REST API for ByteBurrow personal cloud storage",
    "contact": {
      "name": "Martin Miksanik"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/photo": {
      "get": {
        "tags": [
          "photo"
        ],
        "summary": "List all photos without a date\nGET /api/photo",
        "operationId": "list_photos",
        "responses": {
          "200": {
            "description": "List of photos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PhotoResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/photo/list/{year}": {
      "get": {
        "tags": [
          "photo"
        ],
        "summary": "List photos by year\nGET /api/photo/list/:year",
        "operationId": "list_by_year",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "Year",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of photos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PhotoResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid year",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/photo/list/{year}/{month}": {
      "get": {
        "tags": [
          "photo"
        ],
        "summary": "List photos by year and month\nGET /api/photo/list/:year/:month",
        "operationId": "list_by_year_month",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "Year",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "path",
            "description": "Month",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of photos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PhotoResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid year/month",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/photo/list/{year}/{month}/{day}": {
      "get": {
        "tags": [
          "photo"
        ],
        "summary": "List photos by year, month, and day\nGET /api/photo/list/:year/:month/:day",
        "operationId": "list_by_year_month_day",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "Year",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "path",
            "description": "Month",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "day",
            "in": "path",
            "description": "Day",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of photos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PhotoResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/photo/regenerate/{hash}": {
      "post": {
        "tags": [
          "photo"
        ],
        "summary": "Regenerate thumbnail for a photo\nPOST /api/photo/regenerate/:hash",
        "operationId": "regenerate_thumbnail",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Photo hash",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Regeneration queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Photo not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Error response",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "description": "A generic `{ \"message\": \"...\" }` acknowledgement body.\n\nThe single envelope for endpoints whose only useful response is a\nhuman-readable confirmation (deletes, queued jobs, tag updates, …).\nReplaces the scattered ad-hoc `Json(json!({ \"message\": ... }))` so every\nsuch endpoint shares one typed, OpenAPI-documented shape.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "PhotoResponse": {
        "type": "object",
        "required": [
          "hash",
          "keywords"
        ],
        "properties": {
          "date": {
            "type": [
              "string",
              "null"
            ]
          },
          "hash": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "latitude": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "longitude": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          },
          "place": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable location resolved from the EXIF coordinates by the\nreverse-geocoding provider seam (`crate::geo`, #2); `null` until a\nprovider is configured / resolves it."
          },
          "storage_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}